+++ app/utilities/__init__.py
@@ -1,0 +1,9 @@
+
+# Various helper functions
+
+def print_dict(my_dict):
+    for key, value in my_dict.items():
+        print(f'{key}: {value}')
+
+def divider():
+    print('-' * 150)
